+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# websitetoolboxforum Changelog
+# websitetoolboxcommunity Changelog
 
 All notable changes to this project will be documented in this file.
 
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## 1.3.0 - 2021-12-27
 ### Added
-- Website Toolbox forum Settings issue resolved.  
+- Website Toolbox Community Settings issue resolved.  
 
 ## 1.0.0 - 2020-06-30
 ### Added